Fix: Compile issues with syntax changes
str.length() return needed to be cast to int. Fixed warnings from copy
constructor in Player. Makefile now just needs to have objects defined.

+++ Piece.cpp
@@ -58,7 +58,7 @@ void	Piece::draw() {
 }
 
 void	Piece::clear() {
-	for (int i = 0; i < _icon.length(); i++) {
+	for (int i = 0; i < (int)_icon.length(); i++) {
 		mvaddch(_x,_y + i,' ');
 	}
 }
